+++ src/main/java/Task.java
@@ -1,0 +1,29 @@
+public class Task {
+    private final String description;
+    private boolean done = false;
+
+    public Task(String description) {
+        this.description = description;
+    }
+
+    public void mark() {
+        this.done = true;
+    }
+
+    public void unmark() {
+        this.done = false;
+    }
+
+    public boolean isDone() {
+        return this.done;
+    }
+
+    public String getDescription() {
+        return this.description;
+    }
+
+    @Override
+    public String toString() {
+        return "[" + (this.done ? "X" : "") + "]:" + this.description;
+    }
+}
